improve fighting mechanics, change int delay in JSON data structure to float, increase collider size for units, make knight prefab OP for testing purposes

+++ Assets/Scripts/JSON/Spawn.cs
@@ -10,5 +10,5 @@ public class Spawn
     public int startTilePositionIndex;
 
     [JsonProperty("delayBeforeSpawnInSec")]
-    public int delayBeforeSpawnInSec;
+    public float delayBeforeSpawnInSec;
 }
